requirements.txt
alembic
bcrypt
fastapi[all]
password-strength
pre-commit
pydantic-settings
pytest
pytest-asyncio
python-jose[cryptography]
asyncpg
fastapi-users[sqlalchemy]
ruff
slowapi
sqlalchemy[asyncio]
